Trying modifications to TensErLEED code (version 1.61)
Modified the TensErLEED search code to save R-factors per configuration in a hash table, where the list of integers PARIND is the key. Since this requires allocatable arrays, I had to change the search.mpi.f file to .f90, which then requires ViPErLEED to compile it with the '-fixed' tag.

The hash table was implemented with the intarr_hashing.f90 module, modified from fortran_hash_table by Pierre de Buyl; see file header for details.

For benchmarking purposes, I added a parameter TL_VERSION in ViPErLEED, which allows the user to manually select which version of TensErLEED should be used, instead of automatically picking the highest number in the source folder.