Added writeout of hash table contents in TensErLEED 1.61
Added regular writeout of the complete hash table contents as (R-factor, parameters) tuples in TensErLEED. This yields significantly more (and more varied) data than SD.TL, which is important for parabola fits, and potentially for future machine learning. Read-in is currently done by ViPErLEED every 1000 generations; cost of reading the big data files appears to be significant in fast calculations (~ fractions of a second)!

Also fixed a minor issue in TensErLEED: in subroutine Mischung, use NINT instead of INT to round properly.