feat: Allow displacement ranges with even number of steps
Displacement ranges until now only allowed uneven numbers of steps, and would automatically be extended for user input resulting in even numbers of steps. Furthermore, the "centered configurations" produced during the search were done naively by picking the index in the center of the range, which only has the desired result when the range is symmetric around zero.

Ranges with an arbitrary number of steps are now allowed. The "center", i.e. the value closest to zero (no change) of the range is stored in the SearchPar object, and the Rparams.getCenteredConfig now correctly outputs the configuration with least change compared to the known bestfit.