fix: Fixes bug for near-rectangular unit cells
Symmetry search had a higher tolerance for unit cell angles and would therefore sometimes find symmetries for e.g. a rectangular cell; then when the rectangular cell was passed to the guilib part of the program, the lower tolerance there identified the same cell as oblique, and therefore the symmetry as not allowed. Adapted eps to be the same in both program parts.