feat: Adds more information to "Delta Generations" plot in search pdfs
Split the "Generations since last change" plot into separate colors for changes in best, best 10%, and all configurations. Also draw the "max" outline around all three of them. If it is already clear that the max has increased, add another point to the outline at the current generation (i.e. right end of the plot), since this is what's interesting for the convergence criteria.