feat: Makes bookkeeper append contents of a "notes" file to history.info
The bookkeeper will now look for a file "notes" or "notes.txt". If found, the file will be cleared, and the contents will instead be added to the "Notes" section of the added block in the history.info file. This allows recording notes already when the job is being set up, instead of having to run the bookkeeper first and then adding notes afterwards.