feat: Automatically detect BULK_REPEAT vector from N_BULK_LAYERS
If BULK_REPEAT is undefined, the initialization will attempt to detect a bulk repeat vector using the bulk layers, and the same number of layers above. Detection will only work if the layers above are bulk-like (i.e. no relaxation).

If no bulk repeat vector is detected, BULK_REPEAT will default to the previous mode: the repeat vector is assumed to be parallel to the POSCAR c vector, und bulk thickness is taken from the distance to the first non-bulk layer.

In either case, BULK_REPEAT will be written to the PARAMETERS file during initialization to make sure that it stays the same in later runs.