fix: Bulk repeat only needs to match once, not twice
Fixes a minor issue with the detection of the bulk repeat vector introduced previously. The z-range under consideration while checking for translation symmetry was set to twice the thickness of the bulk, which effectively checked whether the bulk and *two* layers above it were translation symmetric for a given vector. Once is enough.