feat: new class LEEDParser
Reads an input file or a dictionary-like structure with similar contents, and handles appropriate conversions of its contents. Also, writes the parameters to file, with an .ini file structure, in which structural domains appear as section headers. Will handle old-style files by adding a dummy section header [S1] (a DeprecationWarning is issued).

Does not chech whether the mandatory parameters are all there. This check is still left to LEEDParameters.