feat: add optional argument 'caller' that is also used for hash computation
This is necessary because two structures with the same basis, same superlattice, but different plane groups would otherwise look the same. Thus the second one was loaded from cache.

One may later get rid of the 'superlattices' key (and parameter) that is only used for computing the hash, as each 'caller' should uniquely identify the structure. Perhaps it would be a good idea to add a __hash__ to LEEDStructuralDomains and LEEDSymmetryDomains that is more advanced than just using self.id().