fix: bug with hash computation gave wrong BeamIndex instances
Very often BeamIndex is instantiated from numerators, thus the input parameters contain integers. Turns out that in python hash(-1) == -2, and hash(obj)  is never == -1, as -1 is used internally in cpython to handle hash errors. Having the cache keys as the hash of a tuple of the input would lead to very strange results when both -1 and -2 were in the input.

Now only BeamIndex instances that do not have a -1 in the input are cached. The others are recomputed every time.