fix: Clear oriState variables after refcalc to force reset
If another search or error calculation is performed after a reference calculation, the "oriState" relative to which offsets are defined should refer to the *new* reference calculation, instead of the previous state. By resetting oriState variables to "None" after a reference calculation, the subsequent delta/search/error sections will automatically store the current state as the new "original" state.