feat: add '--discard' option to bookkeeper
If the bookkeeper is called with the --discard (or -d) option, it will try to delete all output from the last run (including a new Tensor file, if one was generated), not move anything to history, and not write to history.info.

This is meant for cases where e.g. something went wrong with the input, it's clear that the output is worthless, and you don't want to spam the history / history.info

At the moment, this does NOT restore input files, e.g. overwrite PARAMETERS with a PARAMETERS_ori.