hardware: improve error checking on serial read & prevent segfault
rename function decodeMessage() to decodeAndCheckMessage().
* decoding is unchanged
* checks that a message received is acceptable, raises errors otherwise

Prevent possible segmentiation fault when reading from the serial line: we could have written past the end of our buffer array

Added one flag to allow some of the checks: waitingForDataFromPC (bool). It is set in updateState() for the states that require extra data (in a separate message), and reset in those state handlers when the data arrives.