Closes #12
Have readFromSerial() read a bunch of characters in a single state-loop iteration rather than a single one. Characters up to MSG_END are read, waiting 20 us at each character to make sure others that arrive later are also read. Should this not be enough to read in a full message during the currrent state-loop iteration, it will most likely be finished at the end of the next one (if the iteration is slow) or in a few loops if we're doing fast stuff.