hardware: ADC channels should now come as the last bytes in messages
Change order of data bytes for PC_CALIBRATION and PC_SET_UP_ADCS. For both, the ADC channels should come as the last bytes now, not the first ones.

Here is the new sequence of bytes:
- PC_CALIBRATION expects (updateRate, adc0Ch, adc1Ch)
- PC_SET_UP_ADCS expects (#meas MSB, #meas LSB, adc0Ch, adc1Ch)

This sequance has the advantage that it is more extensible (should we ever add more ADCs, the channels can come as successive data bytes).

Documentation updated accordingly