fix: More robust handling of mpirun children in search
In response to a bug report that the search sometimes leaves zombie processes. Not clear why that is, could not reproduce, but one option may be that the pgid cannot be identified any more (because the main subprocess is already dead), but the other children are still alive.
The new solution should take care of this case by storing the pgid right after starting the process, and falling back on that if it cannot be determined any more on the spot.