update: pull master in correctly
For reasons unkonwn, the revert of the last commit did not preserve the 'ahead/behind' commit count with respect to master.

Reverting was necessary as I pulled in some old code from guilib.base that screwed up functioning in here.

The update is now correct. Done by looking at each of the files that changed.