fix: made psgen robust to long paths
psgen passes paths to atom density files to the phaseshift-generation script eeasisss. However, these paths are limited by the fortran line length.

In cases where both the absolute and the relative paths would be too long, instead create a local copy of the relevant atom density files and pass the path to this copy.