fix: Re-write and debug ASA/ASBULK code in refcalc input
ASA and ASBULK are the interlayer bulk vectors. The code for this grew organically with the more advanced ways in which bulk could be defined, and was a) buggy and b) hard to read.

Re-wrote completely, tested on STO and hematite. Both with "float" and "vector" bulk repeat definitions, and with LAYER_STACK_VERTICAL on and off.