feat: Warn if symmetry-equivalent atoms are assigned to different sites
When checking symmetry, check whether atoms at symmetry-equivalent positions have the same site. If not, this could still be intentional, but should output a warning nonetheless.

Note: Technically, site differences could also be checked during the symmetry search, and lower the symmetry in this case. However, there may be usecases where it's useful to have the possibility to break symmetry in occupation/(vibration), but to still keep the geometrical symmetry, and symmetry can always be lowered manually.