feat: scaling of ase cell to experimental bulk
Simple isotropic scaling of the unit cell of the bulk is done before running tleedm. This may be useful to better compare DFT-calculated structures with experimental results. I believe this is the way Lutz does this comparison.

May be good to do from the beginning, especially if we want to train a NN that can work with experimental data directly (although for a NN one would also like to include slightly different bulk lattice constants perhaps?).