feat: origin update when scaling to exp. lattice
Also update origin when rescaling unit cell to fit DFT to experimental lattice parameters. This updates the identification of the z=0 (i.e., highest) atom in the slab. 

This is not updated by default because if the highest atom moves (e.g., +0.1A), then in the subsequent calculations it still has to be treated as being at +0.1 until the next refcalc. So the Cartesian origin is really not consistently the highest atom position, but the original position of the highest atom at the time of the reference calculation.

Related to commit a5be73032dc4db298baf1b946ded0a723b319866