feat: minor modification on handling redundant displacements
If a linked group of atoms gets displacements assigned multiple times (but in a non-conflicting way, e.g. the new list is mirrored with respect to the original one), the default behaviour so far was to overwrite them each time. Now modify the behaviour such that the first assignment is kept unchanged. This removes some redundancy from the code.