Uploading TensErLEED v1.71 as received
Uploading code as received from @LutzHammer

Compatibility with tleedm is a work in progress. Mind that if you place the v1.71 directory in your 'tensorleed' folder, you may need to manually specify TL_VERSION in PARAMETERS to still be able to use tleedm, as it would otherwise try to use the newest version.

Co-Authored-By: LutzHammer <84135857+LutzHammer@users.noreply.github.com>