fix: Rotate hexagonal cells with group cm if necessary
The plane group cm should have the mirror plane oriented along one of the diagonals. However, for hexagonal systems, it is possible to rotate the cell by 30° such that the mirror plane is parallel or perpendicular to one of the two unit vectors.
Added checks for this case, and rotation of the unit cell if required.