feat: Additional parameter in search.steu to disable hash table storage
N-dimensional parabola fits turned out to be rather ineffective, and the hash table storage only improves performance for few-parameter searches that stay unchanged near a local minimum for long times. Since we are now looking at more memory-intensive applications, the tradeoff is usually not worth it, and this makes it possible to disable hash table allocation even within the search. However, this implementation keeps the hash table storage (and output) as an option for future tests.