fix: Jumps in TensErLEED R-factor results
By @LutzHammer
R-factor functions of TensErLEED in both standard rfac and in the search exhibited sudden jumps for some variations, especially clear in error calculations. The problem was an intensity cutoff at 1e-6, below which the beam would not be used; this is relatively common for theoretical beams. When the beam is then "re-activated" because the intensity becomes higher through variation, the R-factor changes drastically.

Lowered cutoff to 1e-8. Note that this generally leads to higher R-factors, as more low-intensity regions are included in the calculation.

Co-Authored-By: LutzHammer <84135857+LutzHammer@users.noreply.github.com>