Removed "Parallelization" from EEASISSS 2021
The new (2021) version of EEASISSS introduced a form of "parallelization" that is incompatible with execution on large clusters (and generally troublesome). This "feature" is now removed and the executables eeasisss and eeas are combined into one. The program EEAS has been transformed into a module which is used directly by EEASISSS rather then executed via the system.
Changes in all four EEASISSS source code files to enable the new logic and workflow.

Additionally, there are changes in psgen.py & recompile_parts.py to work with the changed structure of EEASISSS and account for changed file paths.