fix: last ulog file was not closed properly
File ulogxxx was not closed properly in eeas.f90 which caused an error when trying to access in eeasisss.f90.

The eeasisss code made use of a Fortran peculiarity: Re-opening a different file with the same unit (i.e.open(611,...)) automatically closes the last one. If done in a loop, the next open statement closes the previous – however, the last file will stay open!
This caused a problem with the removal of multithreding, as eeeas.f90 is no longer executed on its own but rather as a part of eeasisss which accesses the same files.

Fixed by adding close(#) statements.

(Also: commented out unused isys calls which stop compilation with ifort)