hardware, camera: trigger burst and roi increments
CameraABC now has an abstract property .supports_trigger_burst that is used to determine whether multiple trigger_now() calls are necessary or not.

Also, minimum increments for ROI width and height are supported. For now, the camera will just complain if the ROI does not fit the minimum increments. We should then have the roi setter functionality (rubber band on live view windw) only allow the supported increments. Probably easier to to directly on the rubber band rectangle.