cleanup, tensorleed: Remove duplicate beam list from input
New TensErLEED version: Previously was reading in 1. a list of indices NPU of output beams to select from the complete beamlist, and 2. an explicit list PQFEX of output beams as h,k couples (subroutine READTL). This was effectively always the same. Simplified by removing READTL and PQFEX. Note that this also changes the format of FIN.

Co-Authored-By: Michele Riva <riva@iap.tuwien.ac.at>