feat, imgsource: safe margin for triggering
When triggering for more than one frame, ask the camera for at least 20% more frames than needed to account for some possible frame loss. This seems to be more than enough in my tests, but could possibly be increased at a later time.

Perhaps one thing that would make it even safer is to decide whether a new trigger burst is to be sent after the current trigger burst is 'over'. This would essentially allow an unlimited number of frames to be averaged.