fix: errors during camera __init__
Should one of the settings be invalid during execution of __init__, errors would not be reported properly, as the error_occurred signal cannot be connected externally before the instance is created.

As a solution, collect errors that may occur during initialization into a list, and report them with a little delay with respect to the end of the __init__ method itself.

Also, make property getters get valid values also in case errors occur. Otherwise there is no way to actually make code run appropriately:
- an error_occurred is emitted
- (hopefully) someone connected it to a handler, that should load new settings. However, before that happens the function that emitted the error has to proceed and finish, and will run with the erroneous value.
- some Exception occurs somewhere else because the value is invalid --> crash; before the fixing implemented externally can ever run.

Probably we will need to do something very similar also on the other abstract base calsses!