Rework of psgen.py
This commit implements a new version of psgen.py to handle input and output for the 2021 version of EEASISS by John Rundgren. Previously, a fixed muffin-tin minimum and maximum radius was passed to the program which would  only work for select systems.

In additions to general improvements, this version contains:
) Completely reworked atom and layer sorting (again).
) rMTmin and rMTmax based on calculation of nearest neighbour distances.
) Proper handling of supercells for element mixing.