feat, camera: pixel intensity limits
New abstract property. Returns min and max value that a pixel can have.

Also, small change in abc: replace thread.quit() with thread.requestInterruption() as I experienced some errors (object in thread was destroyed before .image_saved() could be emitted).
Unclear whether the thread is actually going to be quit with this! To be tested.