hardware: serial bug fix
When switching continuous mode off, we have been discarding data that came after the mode switch. This led to a nasty bug: a part of the three measurements coming from the Arduino was discarded in the process. This screwed over the orer of the data returned.

Fixed by NOT remembering the last command if it is a change_mode command. This will not discard any of the data.