feat: hardwarebase.get_devices
New helper utility that returns all supported and currently available devices for a given 'module' (e.g., 'camera', 'controller'). Relies on the fact that there is a class (driver) for the device, and that this class has a .list_devices method implemented (returning a Sequence of devices).

The devices are returned as a dictionary {name: cls}.