feat: ImageProcessInfo.restore_from
Add method restore_from(other) that just copies the (common) attributes into self. This solves a funny issue with the ImagingSource callback function that would get a wrongly typed object (type 'cell' rather than 'ImageProcessInfo') if the camera process_info is explicitly set to a copy of an ImageProcessInfo instance.