feat: badpixel calculations in other thread
Moved the post-acquisition calculations of bad pixel coordinates and replacement offsets to a separate thread. This has the only purpose of not stalling the UI while calculations are running.

On a side note: it looks like we are using quite a lot of memory to store each indivisual frame. Perhaps it would be better (and faster) to store only the sum of images and the sum of squares (for the variance). This would reduce the memory footprint by a factor of 50 (more or less) for what concerns the acquisition itself. Also, it seems that a spike in memory usage occurs while the calculations are running. Perhaps also worth looking into, after the previous point is taken care of.