fix: invalidate bad pixels info if invalid path
When calling .read() with an invalid path that either does not contain any bad pixels files to read or that contains corrupt files, do not only complain by raising exceptions, but also update the internal info such that the object evaluates to False.

This is useful when picking a new 'invalid' directory after a valid one was chosen, and we catch the exceptions raised.