feat: sizes of uncorrectable pixel clusters
BadPixels now calculates the sizes of uncorrectable bad pixel clusters (if any). The current ROI is used. Uncorrectable pixels that are 4-connected are considered to belong to the same cluster.

The user gets also warned with an explicit pop-up in case there are uncorrectable bad pixel clusters of size larer than 1 (which should always be the case if there are uncorrectable pixesl, except perhaps at edges). This happens only after running a bad-pixels-finding routine.