fix: camera.settings require copying
Make sure that we deepcopy camera.settings whenever we are editing them on the fly. This ensures correct operation of the camera itself. The missing deepcopy would make the check for roi changed fail. This would then not trigger a recalculation of the bad-pixel coordinates onto the new ROI. This means that (i) the bad pixels replacement processing step was wrong (potentially replacing pixels that are not bad), and (ii) it may fail because of bad pixel coordinates overflowing the image size.