fix: binning--roi mismatch
Binning-roi mismatch would trigger a forced replacement of the ROI in the camera. This, however, would commonly lead to other errors, as the modified ROI could be incompatible with the minimum increments in the camera.

New behavior: actually ask the camera for a new ROI only in case the new ROI is a valid one. In any event, during the binning step of image processing we discard the (few) lower-right pixels that do not fit the binning grid.

Also: report the mismatch only once per each connected camera (and per each ROI setting). This prevents annoying spamming of error windows.