fix: serial busy
Previous implementation would not make the serial turn busy if a message did not require a response. Since, however, we rely on the CHANGE of busy in several places, we have to always make the serial busy, and rather forcedly turn it not busy if we do not expect a reply.