feat: emit path to image saved
ImageProcessor.image_saved now carries the full path to the image saved (or an empty string if not image was saved). This is used by CameraABC to itself emit a new image_saved signal, only in case an image was actually saved (while operating in "triggered" mode"). This signal can be used to automatically archive images as soon as they are saved to disk.