fix: NaN values are now recognized in rfactor calculation
Previously, when compiling the new r-factor using aggressive (-Ofast) optimization checks for NaNs would be optimized away and always give the result for a normal number.
This commit adds the compilation flag "-fno-finite-math-only" which enables correct treatment of NaNs.

Additionally, a function "test_compilation" is added to the rfactor module which can be used to test if NaNs are recognized correctly. (This is now also implemented in the ASE API)