feat&fix: treatment of short beams, v0.2.0 of rfactor
This commit adresses a number of conceptual issues in the r-factor calculation.
Firstly, in prepare_beams, it deals with the situation of some beams having less than 2*deg+1 datapoints (which is the minimum required to interpolate them). Those beams will be discarded & filled up with NaNs.
The r-factor calculation can now deal with this case of NaNs in the Y-fuctions.

The ASE API is adapted accordingly to accomodate. The Viperleed implementation will be updated in an upcomming commit.

Further:
)added warning messages

)removed averaging scheme: if required, averaging should be called beforehand
)removed n_beams_out: see above
)removed skip_stages - became unnecessary since none of the stages are skippable now.
)removed old debug statements and deprecated error messages