feat: r-factor v0.2.1: split prepare_beams and y function
Signatures are changed to allow more reuse of arrays without reallocation.
Use new alloc_beams_arrays once in the beginning and then save the allocation time on every single call of the r-factor.
Prepare beams now returns interpolated intensities and derivatives, but not the y function. The y function can be called on these and overwrites the derivative array.