fix: correct interpreation of THEOBEAMS with truncated beams in new rfactor
There was an issue in the new R-factor due to (one of many) peculiarities of TensErLEED:
Beams that contain no data below a threshold energy are reported with intensity of 0.0 instead of NaNs. This was output even into the THEOBEAMS file.
When running ViPErLEED from Tensors the beams are correctly identified as truncated below some value. However, when reading in THEOBEAMS, the beams would be considered to span the entire energy range, but with constant intensity of 0.0 up until the threshold. This, of course, throws of the R-factor massively.