fix: raw frames for bad pixels
Major bug that caused frames to get processed (esp. by removing previously identified bad pixels) while acquiring the frames for detecting bad pixels. This resulted in fewer bad pixels being detected. The main issue is the automatic call to update_bad_pixels() in the bad_pixels getter of CameraABC. This is called any time the underlying BadPixels instance is False-y.

The solution is temporarily invalidating the settings of the camera to an empty bad pixels path. This, in turn, skips the bad pixel removal in the image processor.

Notice that using the frames passed along the frame_ready signal (as opposed to the image_processed signal) is problematic, as non-triggered frames (potentially with incorrect exposure/gain) may get accumulated.

Minor fix: correctly read the previous bad-pixels file from a folder to compare to the new one. This reqquired a signature change for BadPixels.read()