fix: crash opening cam viewer
Attempting to open a camera viewer after running a measurement was failing while loading settings. This seems to be because of an issue stopping the camera (error returned by IC_StopLive). The current (brutal) solution is to completely get rid of the camera and viewer, and recreate them from scratch.