feat: keywords in controller.send_message
ControllerABC.send_message can now take keyword arguments that are passed on unchanged to self.serial.send_message. This is necessary in case one needs to eplicitly pass a timeout that is different than the one in the settings.

Most commonly this would be needed when sending commands that make the controller internally wait before sending a response.

Also: removed the unused make_serial method from ControllerABC