fix: CameraViewer popup after stopped
It would often happen that a CameraViewer with .show_auto == True and .stop_on_close == True would immediately re-appear after a camera had been stopped, as a result of the arrival of some frames (ending up in the event queue). Disconnect appropriate signals in closeEvent to prevent this annoying behavior.