fix: bad pixel removal with roi
Bad pixel removal was not always working appropriately. This had to do with the fact that the Imaging Source camera I tested on (imx265): (i) can only set ROI top-left positions that are multiples of 2, (ii) the camera does not complain if given an odd number and silently sets its lower even multiple, and (iii) the value returned by the get_roi() from the camera is NOT the one actually currently set, but the one requested.

This led to miscalculations in the coordinates of the pixels to be considered as bad.

The current solution changes the return value of get_roi_size_limits(), adding a fourth tuple corresponding to the increments in position. Currently, it is hardcoded as (2, 2) for all Imaging Source cameras. I will need to run some tests with other models to see if this changes. If yes, I'd need to find a way to figure out the correct increments for each camera.