fix: camera n_frames_done
camera.n_frames_done was incorrectly reset every time .trigger_now() was called. This is correct when using trigger burst mode, but would lead to infinite waiting for cameras that need each frame to be triggered via software.

The solution is resetting the number of frames as soon as all frames arrive rather than in trigger_now(). Notice that .n_frames_done is also reset in CameraABC.start() when in triggered mode.