fix: cameras in measurement
Multiple fixes for handling cameras correctly during measurements:
(1) stop timer that triggers cameras as soon as a measurement is aborted. Prevents acquisition of spurious images.
(2) disconnect camera_busy during preparation to prevent misinterpreting a busy=False as a "I have acquired one frame, and we can try going to the next energy". The signal is reconnected once the preparation is over for all devices.
(3) make the connection between camera_busy and _on_camera_busy_changed queued. This is not strictly necassary, but it makes it easier to follow what's going on while debugging.
(4) explcitly avoid trying to go to the next energy if the camera is busy.
(5) use safe_connect to swallow TypeError when connecting camera.stopped to ._finalize. Usually the connection should be unique, but it may not be if we enter _prepare_finalization while aborting because of camera errors.
(6) IV: force cameras to busy at each step, such that we make sure that we will always acquire an image. Otherwise, all controllers may become "not busy" before the camera timer has timed out (i.e., the camera is not busy yet, but has acquired nothing).
(7) Make the camera timer start counting (close to) when the energy has been set (but potentially before I0 is stable)