feat: improve time-resolved intervals
Timer intervals for duration of one "energy step" and for measuring multiple times in triggered mode are now set at runtime rather than at __init__, when stuff may not be yet up to date.

Also, now the duration of a step is bound to the about_to_trigger signal of the primary controller: the countdown starts exactly when the energy is set.

Also, fix: do not try to calculate times for a data point if the measurement was aborted. The recalculation is taken care of already via recalculate_last_step_times(), but we may enter __check_is_finished due to unprocessed timer.timeout events even after aborting.