feat: non-abrupt step profiles
The energy can now be changed in a non-abrupt manner using multiple intermediate steps, as returned by self.step_profile.

This allows, e.g., to speed up measurements as it can reduce the time for settling i0.

Currently implemented profiles:
- 'abrupt' (i.e., no intermediate steps)
- ('linear', n_intermediate_steps, duration)
- list of (energy, interval, energy, interval, ...)

The definition are currently parked in the MeasurementABC class, but will eventually make sense to move them to the EnergyRamp class, once we have one.