feat: speed up SerialABC
To prevent attempting to write to the serial the "next message" while the previous one was still written, we would:
- sent separately each message in a "command, data" sequence
- wait 10ms every time. The communication would take 10ms*n_messages too much.

An improved solution would be to wait 10ms only if the message to be sent is followed by other messages.

However, it looks like a better solution is to combine the multiple messages into a single bytearray, and let the QSerialPort handle the writing. This requires no explicit wait, under the assumption that the .send_message() method will not be called again in the time it takes to write stuff to the serial bus. This should never happen, as the serial is set to busy, unless the command sent does not require a response.