fix: emit error if no data available
We would earlier just silently skip in case one controller had not returned data yet for a completed datapoint.

This can only happen during a time-resolved measurement. The most common reason is that the duration of one energy step has been chosen to be too small.

Typical cases in which this occurs is when the user manipulates the windows: the event loop does not go through for the primary controller while timers expire. This case can probably be solved by moving measurement, datapts, and primary to a separate thread