fix: time delays with multiple controllers
After measuring with two controllers a signal of well-defined frequency (square wave, with a freq.gen.) it turns out that the issue observed at 500Hz sampling is actually on the ACTUAL VALUE OF THE FREQUENCY: the primary controller measures more accurately, whereas the secondary one (that is not stopped, and for which the times are calculated by adding the inverse of the frequency) sees an incorrect frequency.

Allow reading in float values to better correct for frequency mismatches.