fix: crash on closing selector while running
When closing the plugin selector while a measurement was running we would forcibly .quit() the QApplication. This is a problem if some plugins are still doing stuff (e.g., aborting a measurement), as some objects would be suddenly deleted.

Calling QMainwindow.closeEvent() instead, closes the selector window (while the modules close themselves at the appropriate time). The app then terminates since when all windows are closed.