fix: always emit CameraABC.image_saved
the image_saved signal for a camera should be emitted even after the ImageProcessor has been deleted, as long as an image was saved.

The earlier implementation could potentially fail to signal the successful save of an image if a processor was deleted after saving an image and before CameraABC.image_saved could be emitted.

The bug probably came into place now that cameras are faster than earlier.