fix: rework camera .stop()
Replaced the previous implementation of CameraABC.stop() with, essentially, the previous contents of __stop_now_or_retry. The differences are that:
(1) .stop() returns immediately if the camera is not running
(2) it does not emit .stopped() any longer
(3) it returns a boolean, with False value if the camera was not stopped (yet). This may happen either because it was not running, or because it was still waiting for some frames (or for the last image to be processed and saved).

This allows to discern whether anything should be done with the underlying driver: not to be stopped if we're still waiting for frames or if the camera is not running.

The new documentation of .stop() describes how children classes should extend the method.