fix: race condition on next energy step
Fixed one race conditions (observed as occurring):
It was possible for a controller to turn "not busy" while processing the data_ready of another controller and before processing the data_ready of the first one. This could potentially move to the next energy step before all data was collected. Solved by adding a dict of flags that keeps track of which data was already taken care of at the current energy step. Only controllers that can measure are considered. This also saves against the race condition of one controller returning data before the other(s) have even become busy (i.e., would not require to forcedly set each device to busy in non-time-resolved measurements at the beginning of each energy step).