fix: wait for cameras stopped before saving
At the end of the last energy step (or while aborting) we should make sure the cameras are not running before proceeding.

This is necessary because the following call to __disconnect_cameras(), which calls directly camera.disconnect_(), is executed in the thread of the measurement, which may not be the one(s) in which the camera(s) live. This raises QtWarnings about stopping timers in the incorrect thread.

Other minor edit: more general identification of the archive name.