fix: correctly move to another thread
Reimplementation of the moveToThread functionality for all the main base classes. The reimplementation is necessary because of some known bugs in QSerialPort [QTBUG-91237], which are now solved, but only for Qt 6.2, 6.2.5, and 6.5.0. Unfortunately, this would require to use PyQt6 and would break support of Windows 7 platforms, which we cannot afford.

The need for reimplementation of all the moveToThread functionality comes from the fact that, although children are moved together with parents, this does NOT HAPPEN via a call to the child's moveToThread method. This means that only the base-class (i.e., Qt) implementation of moveToThread is automatically called. To do it correctly, reimplementations require to (1) explicitly call child.moveToThread in the parent's moveToThread, and (2) to drop the parent-child relationship. The latter is because QObjects with parent cannot be moved separately from their parent.

As an added complication, the bug on QSerialPort prevents the correct movement of QSerialPort altogether. The only way around this is to RECREATE and REOPEN a QSerialPort IN THE NEW THREAD. This, in turn, implies that the creation and reopening has to be performed by signal-slot, rather than by direct call. This is done in SerialABC via the __move_to_thread_requested signal, connected to __on_moved_to_thread, emitted from the original thread after the previous port is disconnected (in the original thread).

The reimplementation affects moveToThread of
*) MeasurementABC: explicitly call primary_controller.moveToThread; however, we can use the normal parent-child relationship for data_points.
*) ControllerABC: drop parent-child for self.__serial, and explicitly call .serial.moveToThread before moving self.
*) SerialABC: disconnect old serial port, move self and children to thread; (in new thread) create new serial port and connect to it if the old one was connected.

The edits now mean that all the issues previously observed with timers are solved, as start/stop are called in the correct threads now. We can thus revert to use direct calls to __timeout.start/stop.