fix: improve Measure UI code
cleaned up a bit the code of the Measure UI that has accumulated a lot of mess over the last months. Still to do: resort methods

Also, remove all direct function calls to self.measurement methods. Now all is done via signal-slot (was essentially already in place). This is critical for moving to a different thread, as it ensures that methods are executed in the correct thread.