fix: plot and measurement data_points
Currently plot data_points is the same object as measurement.data_points. This is problematic when we will move the measurement (and its data_points) to a different thread, as the object may be accessed by two threads at the same time.

One solution would be to use mutexes and locking. This would however probably slow down the measurement, and hold its thread locked while stuff happens on the UI (e.g., moving windows around). The better option seems to have two copies of the DataPoints object, kept in sync in the GUI thread via the new_data_available signal. For this reason, I have changed the signature of the new_data_available signal to carry the new data point (a dict) just acquired.