feat: CameraABC.extra_delay
Add a new abstract property that should return the time interval, in addition to the one expected from the settings, that the camera may take to become not busy after a trigger_now. This can be used to judge how long a camera requires, and suggest optimal configurations later on.