fix: broken triggered time-resolved
The introduction of __data_stored() in commit 4a1ab5583eb1ba2df42840ecfd8e880d142482c1 was not an issue for "continuous" (i.e., spamming) time-resolved measurements because we would fully reimplement _ready_for_next_measurement there. However, it is a problem for a triggered type of measurement (which uses the base-class implementation of _ready_for_next_measurement), as _data_stored() was always containing False (because of the reimplemented _on_controller_data_ready).

The solution is (i) making __data_stored a protected but not private attribute, and (ii) always replace it with an empty dictionary in time-resolved measurements.