fix, feat: begin triggered time-resolved
In the previous implementation, we would not be measuring for the first measurement_interval milliseconds of each energy step. This is because we would .start the __trigger_one_measurement timer at the beginning of an energy step. The better solution is to always ask for measurements to be take via the set_leed_energy command, and rather start the timer via .about_to_trigger. This way, the first measurement is taken after the energy is set, and from then on more measurements are acquired at the right intervals.

This modification affects the minimum duration of an energy step (becomes shorter, just a bit longer than one measurement_interval instead of two).

Also: catch AttributeError exceptions in method that may get called before attributes exist.

Notice that the TIMES are still INCORRECT for a triggered measurement, because of a bug in DataPoints