feat: DataPoints.continuous attribute
New attribute that can be given at instantiation or set later via appropriate setters. It is such that one can never have a (non-time_resolved and continuous).

The presence of the attribute allows a simplification of the code of TimeResolved, in that __check_is_finished does exactly the same as _ready_for_next_measurement.

The attribute is also set correctly when reading from folder/archive, by taking the information from the measurement config.