fix: rework DataPoint time axis calculation
Completely rework the determination of the time axis for data. This was needed to correctly handle triggered time-resolved measurements, and makes the code more readable.

(1) Introduce a new QuantityInfo.TIMESTAMPS that is only kept internally and never stored to file.
(2) store a QuantityInfo.TIMESTAMPS with every data point of each controller in energy-resolved and triggered time-resolved modes. Only one per energy is necessary in continuous time-resolved. These are automatically included in ControllerABC.
(3) Calculate the time axis using these TIMESTAMPS, that are not overwritten now (as opposed to the TIMES in the previous implementation).

This solution allows us to also get rid of the recalculation of the time axis as a separate method. It also allows to remove a bunch of obscure private attributes.