feat: t-resolved multiple images per energy
Allow multiple images to be acquired in triggered time-resolved mode. This requires to:

(1) add image names once they are fully acquired (but perhaps not processed and saved yet) rather than at the beginning of an energy step.
(2) have DataPoints contain a list of IMAGES rather than a single one. When saving to file, one has to account for the fact that the number of images may not match the number of measurements; i.e., they have to be "justified" the same way as other quantities.
(3) Acquiring the images should be done at the same time as measuring the other quantities. The simplest way is to use the ._camera_timer (already connected appropriately in ABC), set its timeout to 0, and connect whatever signal triggers measurement to its .start slot. This may introduce a slight misalignment between measurements and images if the event loop gets stuck. This should not be an issue, though, as there is nothing freezing the GUI event loop.
(4) The time needed by the cameras to return images has to be taken into consideration when determining the minimum interval between "measurements".